+++ src/App.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
-import TodoList from "./features/TodoList/TodoList";
-import TodosViewForm from "./features/TodoList/TodosViewForm";
+import TodoList from "./features/TodoList";
+import TodosViewForm from "./features/TodosViewForm";
+import TodoForm from "./TodoList/TodoForm";
 
 const encodeUrl = ({ sortField, sortDirection, searchQuery, url }) => {
   let sortQuery = `sort[0][field]=${sortField}&sort[0][direction]=${sortDirection}`;
@@ -259,34 +260,13 @@ const TodoApp = () => {
           style={{ width: "100%", padding: "0.25rem" }}
         />
       </div>
-
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          const input = e.target.elements.newTodo;
-          const value = input.value.trim();
-          if (value) {
-            handleAddTodo({ title: value });
-            input.value = "";
-          }
-        }}
-        style={{ marginBottom: "1rem" }}
-      >
-        <input
-          name="newTodo"
-          placeholder="New todo"
-          style={{ marginRight: "0.5rem" }}
-        />
-        <button type="submit" disabled={isSaving}>
-          {isSaving ? "Saving..." : "Add"}
-        </button>
-      </form>
+      <TodoForm handleAddTodo={handleAddTodo} isSaving={isSaving} />
 
       <TodoList
         todos={todoList}
+        onComplete={completeTodo}
+        onUpdate={updateTodo}
         isLoading={isLoading}
-        onToggleComplete={completeTodo}
-        onUpdateTodo={updateTodo}
       />
 
       <hr />
